# openapi/mailbox_v0.1.yaml
openapi: 3.0.3
info:
  title: WHISPER Mailbox API
  version: "0.1"
servers:
  - url: https://example-mailbox-server

paths:
  /v1/mailboxes:
    post:
      summary: Create a mailbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMailboxRequest"
      responses:
        "200":
          description: Mailbox created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMailboxResponse"

  /v1/mailboxes/{mailbox_id}/deposit-tokens:
    post:
      summary: Register deposit tokens for a mailbox (owner only)
      parameters:
        - $ref: "#/components/parameters/MailboxId"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDepositTokensRequest"
      responses:
        "200":
          description: Tokens registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterDepositTokensResponse"

  /v1/mailboxes/{mailbox_id}/deposit:
    post:
      summary: Deposit an encrypted blob into a mailbox
      parameters:
        - $ref: "#/components/parameters/MailboxId"
        - name: X-Whisper-MsgId
          in: header
          required: true
          schema:
            type: string
          description: base64url(16..32 bytes)
        - name: X-Whisper-ExpiresAt
          in: header
          required: false
          schema:
            type: integer
          description: unix timestamp (optional)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
        "409":
          description: Duplicate msg_id

  /v1/mailboxes/{mailbox_id}/poll:
    get:
      summary: Poll messages from a mailbox
      parameters:
        - $ref: "#/components/parameters/MailboxId"
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResponse"

  /v1/mailboxes/{mailbox_id}/ack:
    post:
      summary: Acknowledge (delete) messages by msg_id
      parameters:
        - $ref: "#/components/parameters/MailboxId"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckRequest"
      responses:
        "200":
          description: Deleted count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"

  /v1/mailboxes/{mailbox_id}/revoke:
    post:
      summary: Revoke deposit tokens by hashed values (owner only)
      parameters:
        - $ref: "#/components/parameters/MailboxId"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
      responses:
        "200":
          description: Revoked count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  parameters:
    MailboxId:
      name: mailbox_id
      in: path
      required: true
      schema:
        type: string

  schemas:
    CreateMailboxRequest:
      type: object
      properties:
        poll_token:
          type: string
          nullable: true
          description: base64url(32 bytes), if client-generated
      additionalProperties: false

    Limits:
      type: object
      properties:
        max_msg_bytes: { type: integer }
        max_queue_bytes: { type: integer }
        ttl_days: { type: integer }
      required: [max_msg_bytes, max_queue_bytes, ttl_days]

    CreateMailboxResponse:
      type: object
      properties:
        mailbox_id: { type: string }
        poll_token:
          type: string
          nullable: true
          description: Present only if server-generated
        limits:
          $ref: "#/components/schemas/Limits"
      required: [mailbox_id, limits]

    RegisterDepositTokensRequest:
      type: object
      properties:
        deposit_tokens:
          type: array
          items: { type: string }
          description: base64url(32 bytes) tokens
      required: [deposit_tokens]

    RegisterDepositTokensResponse:
      type: object
      properties:
        added: { type: integer }
      required: [added]

    DepositResponse:
      type: object
      properties:
        stored: { type: boolean }
        msg_id: { type: string }
        expires_at: { type: integer }
      required: [stored, msg_id, expires_at]

    PollMessage:
      type: object
      properties:
        msg_id: { type: string }
        received_at: { type: integer }
        expires_at: { type: integer }
        blob_b64: { type: string }
      required: [msg_id, received_at, expires_at, blob_b64]

    PollResponse:
      type: object
      properties:
        cursor: { type: string }
        messages:
          type: array
          items:
            $ref: "#/components/schemas/PollMessage"
      required: [cursor, messages]

    AckRequest:
      type: object
      properties:
        msg_ids:
          type: array
          items: { type: string }
          description: base64url msg ids
      required: [msg_ids]

    AckResponse:
      type: object
      properties:
        deleted: { type: integer }
      required: [deleted]

    RevokeRequest:
      type: object
      properties:
        deposit_token_hashes:
          type: array
          items: { type: string }
          description: base64url(32 bytes) dep_hash values
      required: [deposit_token_hashes]

    RevokeResponse:
      type: object
      properties:
        revoked: { type: integer }
      required: [revoked]